renamed component ItemsNavPrevNext
